+++ Pocketchi/src/utils/helpers/Texts.h
@@ -4,13 +4,13 @@ struct Texts
 {
   Tinyfont tinyfont = Tinyfont(Arduboy2::sBuffer, Arduboy2::width(), Arduboy2::height());
 
-  void printLine(uint8_t x, uint8_t y, char *line)
+  void printLine(size_t x, size_t y, char *line)
   {
     tinyfont.setCursor(x, y);
     tinyfont.print(line);
   }
 
-  void printValue(uint8_t x, uint8_t y, size_t value)
+  void printValue(size_t x, size_t y, size_t value)
   {
     tinyfont.setCursor(x, y);
     tinyfont.print(value);
